+++ app/Http/Controllers/UserManagementController.php
@@ -52,37 +52,37 @@ class UserManagementController extends Controller
         ]);
     }
 
-    public function userRegister(Request $request)
-    {
-        $role = $request->role;
-        $this->authorize($role === 'Etudiant' ? 'create_student' : 'create_pilot');
+        public function userRegister(Request $request)
+        {
+            $role = $request->role;
+            $this->authorize($role === 'Etudiant' ? 'create_student' : 'create_pilot');
 
-        $data = $request->validate([
-            'name' => 'required|string|max:255',
-            'first_name' => 'required|string|max:50',
-            'email' => 'required|email|unique:users|max:255',
-            'password' => 'required|string|min:6|confirmed|regex:/^(?=.*[A-Z])(?=.*[a-z])(?=.*\d).+$/',
-            'pp' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
-            'birthdate' => 'required|date|before:today|after:1900-01-01',
-            'city_id' => 'required|exists:cities,id',
-            'classe_id' => 'nullable|exists:classes,id',
-            'new_classe' => 'nullable|string|max:50|unique:classes,name',
-        ]);
+            $data = $request->validate([
+                'name' => 'required|string|max:255',
+                'first_name' => 'required|string|max:50',
+                'email' => 'required|email|unique:users|max:255',
+                'password' => 'required|string|min:6|confirmed|regex:/^(?=.*[A-Z])(?=.*[a-z])(?=.*\d).+$/',
+                'pp' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
+                'birthdate' => 'required|date|before:today|after:1900-01-01',
+                'city_id' => 'required|exists:cities,id',
+                'classe_id' => 'nullable|exists:classes,id',
+                'new_classe' => 'nullable|string|max:50|unique:classes,name',
+            ]);
 
-        $classe = $this->getOrCreateClasse($request);
-        $data['classe_id'] = $role === 'Etudiant' ? optional($classe)->id : null;
-        $data['pp_path'] = $request->file('pp') ? $request->file('pp')->store('images', 'public') : 'images/profile_picture.jpg';
-        $data['password'] = Hash::make($data['password']);
+            $classe = $this->getOrCreateClasse($request);
+            $data['classe_id'] = $role === 'Etudiant' ? optional($classe)->id : null;
+            $data['pp_path'] = $request->file('pp') ? $request->file('pp')->store('images', 'public') : 'images/profile_picture.jpg';
+            $data['password'] = Hash::make($data['password']);
 
-        $user = User::create($data);
-        $user->assignRole($role);
+            $user = User::create($data);
+            $user->assignRole($role);
 
-        if ($role === 'Pilote' && $request->has('classesPilots')) {
-            $user->classesPilots()->attach($request->classesPilots);
+            if ($role === 'Pilote' && $request->has('classesPilots')) {
+                $user->classesPilots()->attach($request->classesPilots);
+            }
+
+            return redirect()->route($role === 'Etudiant' ? 'students_list' : 'pilots_list');
         }
-
-        return redirect()->route($role === 'Etudiant' ? 'students_list' : 'pilots_list');
-    }
 
     public function showUserInfo($role, $id)
     {
@@ -120,7 +120,7 @@ class UserManagementController extends Controller
     {
         $user = User::findOrFail($id);
         $role = $user->hasRole('Etudiant') ? 'Etudiant' : 'Pilote';
-        $this->authorize($role === 'Etudiant' ? 'edit_student' : 'edit_pilot');
+        $this->authorize(ability: $role === 'Etudiant' ? 'edit_student' : 'edit_pilot');
 
         $data = $request->validate([
             'name' => 'required|string|max:255',
@@ -151,8 +151,12 @@ class UserManagementController extends Controller
 
     public function deleteUser($id)
     {
+        
         $user = User::findOrFail($id);
         $role = $user->hasRole('Etudiant') ? 'Etudiant' : 'Pilote';
+
+        $this->authorize(ability: $role === 'Etudiant' ? 'delete_student' : 'delete_pilot');
+
         // On effectue une suppression douce
         $user->delete();
